add casts to get Mac builds to build for CVS: ----------------------------------------------------------------------
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@2848 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/Network/network_star_hub.cpp
@@ -1163,7 +1163,7 @@ send_packets()
 				int32 startTick;
 				int32 endTick;
 				// never send fewer than 2 full updates per second, or more than 15
-				int effectiveLatency = std::min(2, std::max(thePlayer.mDisplayLatencyTicks / (int32) thePlayer.mDisplayLatencyBuffer.size(), TICKS_PER_SECOND / 2));
+				int effectiveLatency = std::min((int32) 2, std::max((int32) (thePlayer.mDisplayLatencyTicks / thePlayer.mDisplayLatencyBuffer.size()), (int32) (TICKS_PER_SECOND / 2)));
 				if (1) {
 					if (sNetworkTicker - thePlayer.mLastRecoverySend >= effectiveLatency)
 					{
